agent_worker/requirements.txt
livekit-agents
python-dotenv
livekit-plugins-openai
livekit-plugins-noise-cancellation
livekit-plugins-silero
livekit
openai